refactor(g1): rename Effects tab to ItemEffects for naming parity
