+++ docs/migration-plan/poc/refactor-naming-conventions.md
@@ -584,7 +584,7 @@ Every numbered task ends with `npm run build` exit 0. Failures block the next ta
 ### Final validation
 - [ ] FV.1 — Workspace `grep_search` for old prefix forms (`DnD35e`, `BaseDnd35eSystemData`, `BaseDnd35eSystemSource`, `ItemSystemModelBase`, `ActorSystemModelBase`, `ActiveEffectSystemModelBase`, `PhysicalSystemData.mjs`, `PhysicalItemDnd35e`, `BaseItemSheet`, `applyIdentifiableSchema.mjs`, `Dnd35eDocument` class, `Dnd35eDocumentMixin`, `Dnd35eDocumentSystemModel`, `Dnd35eDocumentFlags`, `Dnd35eDocumentProperties`, `Dnd35eDocumentConstructor`, `Dnd35eActiveEffect` class, `Dnd35eActiveEffectFlags`, `Dnd35eActiveEffectConfig`, `Dnd35eActiveEffectSystemSource`, `Dnd35eActiveEffectSource`, `Dnd35eParentDoc`, `Dnd35eFieldMeta`, `Dnd35eOverrideOptions`, `Dnd35eSectionField`, `Dnd35eSystemConfig`, `Dnd35eBaseFlags`, `Dnd35eEffectChangeData`, `Dnd35eChangeType`, `DND35E_CHANGE_TYPE`, `tokenDnd35e` (lowercase class form), `logHelper.mjs`, `src/scene/`, `region-document`, `token-document`, `region-behaviour`, `@settings/_types`, `@settings/constants`, `_types.mjs`, `_types.mts`, `@ec/`, `@entities/`, `src/entities/`, inline `getSchemaField` definitions, inline `syncOpenSheetTitle` definitions, top-level `BaseActiveEffect/resolveChangeValue.mjs`). Production code must show zero matches; planning docs / changelog allowed.
 - [ ] FV.2 — `npm run build` clean
-- [ ] FV.3 — `npx vitest run --project unit` green
+- [ ] FV.3 — `npx vitest run` green
 - [ ] FV.4 — **Full Playwright E2E suite** (`npx playwright test`) green on the final merged `dev` tree (one last sanity pass; each PR already ran E2E pre-merge).
 - [ ] FV.5 — `.github/copilot-instructions.md` and `.github/instructions/dnd35e-patterns.instructions.md` updated to reflect the new `documents/` vocabulary, the suffix rule (with verified-collision exceptions), the `Base`-over-`Dnd35e` preference, the file-name-vs-grouped-helpers exception, the R.2.9 file-casing rule, and the R.2.10 `types.mts` convention.
 - [ ] FV.6 — Mark this phase doc Complete (no roadmap row needed — see R.7).
