+++ src/entities/items/baseItem/sheet/tabs/ItemEffects.vue
