+++ src/entities/items/baseItem/sheet/tabs/index.mts
@@ -1,16 +1,16 @@
 import type { SheetTab } from '@ec/CoreMixin/index.mjs';
 
-import Effects from './Effects.vue';
+import ItemEffects from './ItemEffects.vue';
 
 const defaultEffectsTab: SheetTab = {
   id: 'effects',
   label: 'dnd35e.EFFECT.Effects',
-  component: Effects,
+  component: ItemEffects,
   order: 100,
   icon: 'fas fa-bolt',
 };
 
 export {
   defaultEffectsTab,
-  Effects,
+  ItemEffects,
 };
