+++ src/entities/items/components/Physical/sheet/tabs/PhysicalItemEffects.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
   import { DocumentSheetStoreSymbol } from '@ec/CoreMixin/index.mjs';
-  import { Effects } from '@items/baseItem/index.mjs';
+  import { ItemEffects } from '@items/baseItem/index.mjs';
   import { inject } from 'vue';
 
   import MaterialsList from '../components/MaterialsList.vue';
@@ -17,10 +17,10 @@
 </script>
 
 <template>
-  <Effects :has-added-effects="!!materials.length || !!secrets.length">
+  <ItemEffects :has-added-effects="!!materials.length || !!secrets.length">
     <template #effects-list-prepend>
       <SecretsList />
       <MaterialsList />
     </template>
-  </Effects>
+  </ItemEffects>
 </template>
